Update FlexRegBuilder.java
Made it so the Create button will not overwrite already existing files. Will output an error instead.
Tweaked the add folder option so that it will give no error by default, but will give a warning if at least one file in folder is not .xml. NOTE: It's okay if some files are not .xml, they will just be ignored.